refactor(upgrade): Update NgMetadataUpgradeAdapter to support NgModule
BREAKING CHANGE: NgMetadataUpgradeAdapter now accepts an already
instantiated @angular/upgrade UpgradeAdapter, which will have been
created using an Angular 2 NgModule.

**Before:**
```ts
import { NgMetadataUpgradeAdapter } from 'ng-metadata/upgrade'
import { UpgradeAdapter } from '@angular/upgrade'
import { Component } from 'ng-metadata/core'

const angular1Module = angular.module('ng1Module', [])

@Component({
  selector: 'app'
})
class AppComponent {}

const upgradeAdapter = new NgMetadataUpgradeAdapter(UpgradeAdapter)

upgradeAdapter.bootstrap(AppComponent, ['ng1Module'])
```

**After:**
```ts
import { NgMetadataUpgradeAdapter } from 'ng-metadata/upgrade'
import { UpgradeAdapter } from '@angular/upgrade'
import { NgModule } from '@angular/core'

const angular1Module = angular.module('ng1Module', [])

@NgModule({
  selector: 'ng2'
})
class Ng2Module {}

const upgradeAdapter = new NgMetadataUpgradeAdapter( new
UpgradeAdapter(Ng2Module) )

upgradeAdapter.boostrap(document.body, ['ng1Module'])
```
